refactor: move deal.constants to deal folder
