+++ components/users/user-home.constants.ts
@@ -1,4 +1,4 @@
-import { DEAL_OPTIONS } from '../deals/deal.constants';
+import { DEAL_OPTIONS } from '../../lib/deal/deal.constants';
 
 export const USER_HOME_OPTIONS = [
   { value: 'home', label: '홈' },
