+++ lib/deal/deal.constants.ts
