+++ lib/deal/deal.types.ts
@@ -1,0 +1,1 @@
+export type Deal = 'offer' | 'demand' | 'swap';
